Create shared capture discover hook and basic wiring to rediscover cta
