+++ src/components/editor/Bindings/Selector.tsx
@@ -3,8 +3,10 @@ import ErrorOutlineIcon from '@mui/icons-material/ErrorOutline';
 import {
     Box,
     Button,
+    Divider,
     IconButton,
     ListItemText,
+    Stack,
     Typography,
 } from '@mui/material';
 import {
@@ -45,6 +47,7 @@ interface BindingSelectorProps {
     loading: boolean;
     skeleton: ReactNode;
     readOnly?: boolean;
+    RediscoverButton?: ReactNode;
 }
 
 interface RowProps {
@@ -118,6 +121,7 @@ function BindingSelector({
     loading,
     skeleton,
     readOnly,
+    RediscoverButton,
 }: BindingSelectorProps) {
     const onSelectTimeOut = useRef<number | null>(null);
 
@@ -165,18 +169,7 @@ function BindingSelector({
             headerName: collectionsLabel,
             sortable: false,
             renderHeader: (params: GridColumnHeaderParams) => (
-                <>
-                    <Typography>{params.colDef.headerName}</Typography>
-
-                    <Button
-                        variant="text"
-                        disabled={formActive}
-                        onClick={handlers.removeAllCollections}
-                        sx={{ borderRadius: 0 }}
-                    >
-                        <FormattedMessage id="workflows.collectionSelector.cta.delete" />
-                    </Button>
-                </>
+                <Typography>{params.colDef.headerName}</Typography>
             ),
             renderCell: (params: GridRenderCellParams) => {
                 const currentConfig = resourceConfig[params.row];
@@ -221,6 +214,35 @@ function BindingSelector({
     ) : (
         <>
             <CollectionPicker readOnly={readOnly} />
+
+            <Box
+                sx={{
+                    marginLeft: 'auto',
+                }}
+            >
+                <Stack
+                    direction="row"
+                    spacing={1}
+                    divider={
+                        <Divider
+                            orientation="vertical"
+                            variant="middle"
+                            flexItem
+                        />
+                    }
+                >
+                    {RediscoverButton ?? null}
+
+                    <Button
+                        variant="text"
+                        disabled={formActive}
+                        onClick={handlers.removeAllCollections}
+                        sx={{ borderRadius: 0 }}
+                    >
+                        <FormattedMessage id="workflows.collectionSelector.cta.delete" />
+                    </Button>
+                </Stack>
+            </Box>
 
             <Box sx={{ height: 280 }}>
                 <DataGrid
